+++ src/api/iscomputed.ts
@@ -1,8 +1,9 @@
 import { isObservableObject } from "../types/observableobject"
 import { getAtom } from "../types/type-utils"
 import { isComputedValue } from "../core/computedvalue"
+import { fail } from "../utils/utils"
 
-export function isComputed(value, property?: string): boolean {
+export function _isComputed(value, property?: string): boolean {
     if (value === null || value === undefined) return false
     if (property !== undefined) {
         if (isObservableObject(value) === false) return false
@@ -11,4 +12,18 @@ export function isComputed(value, property?: string): boolean {
         return isComputedValue(atom)
     }
     return isComputedValue(value)
+}
+
+export function isComputed(value: any): boolean {
+    if (arguments.length > 1)
+        return fail(
+            `isComputed expects only 1 argument. Use isObsevableProp to inspect the observability of a property`
+        )
+    return _isComputed(value)
+}
+
+export function isComputedProp(value: any, propName: string): boolean {
+    if (typeof propName !== "string")
+        return fail(`isComputed expected a property name as second argument`)
+    return _isComputed(value, propName)
 }